requirements.txt
openai
whisper
pinecone-client
langchain
transformers
torch
pandas
scikit-learn
